incorporate basic worker fsm
The worker fsm is designed to prevent the worker from
either getting caught in a loop or accepting more tasks
than it should at any given time.

The goal is to also allow Worker(s) to remain in a suspended
state once something goes wrong, delegating control to the
supervisor to determine if it should restart the worker or
not, for the same task.

It's still very much wip, there's a lot I don't like about
this design, but the client interface is getting closer to
what I'm expecting it to be -- a simple implementation of
a worker against a task with some context, which you wrap
in a `Worker` such that it can be input as part of the
supervisor pool.