misc: refactor supervisor shutdown
rather than relying on joinset/joinhandle we emit
`Shutdown` messages and wait for `ShutdownAck`
acknowledgements from all workers (both pooled and
checked workers).

This is done in anticipation of supporting workers
that are not on the same binary as the supervisor
(i.e., running as tasks) but can be elsewhere, such
as a different network.