misc(supervisor): refactor supervisor struct
Refactored `pool` and `checked` queues from `VecDequeue` to
`BTreeeMap`. This simplifies lookups while still allowing the
flexibility to pop from the map.

Added `tasks` as the internal tasks queue. When receiving a task
request we now simply push it to the back of the queue rather
than process it immediately. This brings a number of benefits:

* Ensures task ordering is maintained, where pending tasks do not
  remain pending as long as new tasks are being enqueued;
* Allows us to add timer/notify on the task queue whenever a worker
  is available to then process the task.