misc: adding `BoundedBTreeMap` and cleaning up spawning
Still WIP, but wanted to introduce bounded maps to ensure
that thereis no way to exceed the pool limit. The design
question that we have right now is that we have the `pool`
containing all the workers ready to take a task and the
`checked` with all the workers currently in a non-`Idle`
state. This means that we _still_ need to perform bound
checks when spawning across both maps.

One solution may be to merge the two into a single map and
add a flag to the `WorkerCtx` indicating if a worker is
currently checked in or not.